+++ gen/com/yanncebron/m68kplugin/lang/psi/M68kMoveaInstruction.java
@@ -21,29 +21,29 @@ import com.intellij.psi.PsiElement;
 
 public interface M68kMoveaInstruction extends M68kMoveInstructionBase {
 
-  @Nullable
-  M68kAdmAbs getAdmAbs();
+  @NotNull
+  List<M68kAdmAbs> getAdmAbsList();
 
-  @Nullable
-  M68kAdmAdi getAdmAdi();
+  @NotNull
+  List<M68kAdmAdi> getAdmAdiList();
 
-  @Nullable
-  M68kAdmAix getAdmAix();
+  @NotNull
+  List<M68kAdmAix> getAdmAixList();
 
-  @Nullable
-  M68kAdmApd getAdmApd();
+  @NotNull
+  List<M68kAdmApd> getAdmApdList();
 
-  @Nullable
-  M68kAdmApi getAdmApi();
+  @NotNull
+  List<M68kAdmApi> getAdmApiList();
 
   @NotNull
   List<M68kAdmArd> getAdmArdList();
 
-  @Nullable
-  M68kAdmAri getAdmAri();
+  @NotNull
+  List<M68kAdmAri> getAdmAriList();
 
-  @Nullable
-  M68kAdmDrd getAdmDrd();
+  @NotNull
+  List<M68kAdmDrd> getAdmDrdList();
 
   @Nullable
   M68kAdmImm getAdmImm();